move Menu to standalone module
